% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outcome.R
\name{assign_outcome}
\alias{assign_outcome}
\title{Assign outcomes (i.e., hit, miss, correct rejection, false alarm)}
\usage{
assign_outcome(data, acc_var = "accuracy", signal_var = "signal")
}
\arguments{
\item{acc_var}{Quoted column name coding for accuracy.
Accuracy data should be binary with 1 coding for correct and 0 incorrect.}

\item{signal_var}{Quoted column name indicating signal status
(e.g., old/new, change/no change). Signal data should be binary with 1
coding for signal present and 0 coding for signal absent}

\item{id_var}{Quoted column name coding for participant ID.}
}
\description{
Function called by the user to assign outcomes.
Data should include a column coding for accuracy and a column coding for
signal type
}
\examples{
# using assign_outcome() with change detection data
outcome_data <- assign_outcome(my_data,
                               acc_var = "accuracy",
                               signal_var = "change_type")

}
