% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_measures.R
\name{calc_measures}
\alias{calc_measures}
\title{Calculate various SDT measures.}
\usage{
calc_measures(data, hit_var = "p_hit", fa_var = "p_fa", measure = NULL)
}
\arguments{
\item{hit_var}{Quoted column name coding for proportion of hits.}

\item{fa_var}{Quoted column name coding for proportion of false alarms.}

\item{measure}{A string indicating which measure to calculate. At present,
only d' and criterion measures are implemented.}
}
\description{
Function called by user to calculate various SDT measures. Data should
include columns coding for proportion of hits and proportion of false
alarms. Data should also include outcome totals (i.e., columns titled
"n_hit", "n_miss", "n_cr", "n_fa") returned from get_props() function. This
allows automatic correction of proportions and recalculation of called
measure should extreme values be detected.
}
\examples{
\dontrun{
# using calc_measures() with proportions data from get_props() example,
# passing d' as the measure to calculate
d_data <- calc_measures(proportions,
                        hit_var = "p_hit",
                        fa_var = "p_fa",
                        measure = "d_prime")

# using calc_measures() with proportions data from get_props() example,
# passing criterion as the measure to calculate
c_data <- calc_measures(proportions,
                        hit_var = "p_hit",
                        fa_var = "p_fa",
                        measure = "criterion")
}
}
