% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Calculate criterion.}
\usage{
criterion(data, id_var = "id", hit_var = "p_hit", fa_var = "p_fa")
}
\arguments{
\item{id_var}{Quoted column name coding for participant ID.}

\item{hit_var}{Quoted column name coding for porportion of hits.}

\item{fa_var}{Quoted column name coding for proportion of false alarms.}
}
\description{
Function called by user to calculate criterion. Data should include columns coding
for proportion of hits and proportion of false alarms.
}
