% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probs.R
\name{get_probs}
\alias{get_probs}
\title{Assign conditional probabilites}
\usage{
get_probs(
  data,
  id_var = "id",
  accuracy_var = "accuracy",
  signal_var = "signal_type",
  condition_var = NULL
)
}
\arguments{
\item{id_var}{Quoted column name coding for participant ID.}

\item{signal_var}{Quoted column name coding for condition variable.
Condition data should be binary with 1 coding for signal present and
0 for signal absent (e.g., change/no change; old/new).}

\item{acc_var}{Quoted column name coding for accuracy.
Accuracy data should be binary with 1 coding for correct and 0 incorrect.}
}
\description{
Function called by the user to assign conditional probabilities to data.
Data should include accuracy (binary) and a signal variable (e.g., change/no change).
}
