% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_measures.R
\name{plot_measures}
\alias{plot_measures}
\title{Plot various SDT measures.}
\usage{
plot_measures(data, id_var = "id", measure = NULL, condition_var = NULL)
}
\arguments{
\item{id_var}{Quoted column name coding for participant ID.}

\item{measure_var}{Quoted column name coding for measure values.}
}
\description{
Function called by user to plot various SDT measures. Data should include
columns containing measure values calculated by the calc_measures() function
as well as a column containing participant ID.
}
\examples{
\dontrun{
# using plot_measures() with d' data obtained from calc_measures() function
d_plot <- plot_measures(d_data,
                        id_var = "id",
                        measure = "d_prime",
                        condition_var = NULL)
}
}
