% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Calculate proportion of hits, misses, correct rejections, and false alarms.}
\usage{
get_props(
  data,
  id_var = "id",
  probability_var = "con_p",
  return = "raw",
  condition_var = NULL
)
}
\arguments{
\item{id_var}{Quoted column name coding for participant ID.}

\item{return}{If set to "proportions", returns a data frame
containing the proportions of each conditional probability. If set to
"raw" (default), returns the proportions of each conditional probability as well
as the total number of trials for each conditional probability.}

\item{conditional_var}{Quoted column name coding for conditional
probability for each row.}
}
\description{
Function called by the user to calculate proportions of each conditional
probability. Data should include a column containing the conditional
probabilities assigned using the "get_probs" function.
}
