% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_props.R
\name{get_props}
\alias{get_props}
\title{Calculate proportion of hits, misses, correct rejections, and false alarms
and return totals for each outcome.}
\usage{
get_props(
  data,
  id_var = "id",
  outcome_var = "outcome",
  condition_var = NULL,
  correction = NULL
)
}
\arguments{
\item{id_var}{Quoted column name coding for participant ID.}

\item{outcome_var}{Quoted column name coding for outcomes assigned by the
assign_outcome() function.}

\item{condition_var}{Quoted column name coding for condition.}
}
\description{
Function called by the user to calculate proportions of each outcome.
Data should include a column containing the outcomes assigned using the
assign_outcome() function as well as a column containing participant ID.
}
